+++ lib/src/service/tweets/media.freezed.dart
@@ -1,0 +1,171 @@
+// coverage:ignore-file
+// GENERATED CODE - DO NOT MODIFY BY HAND
+// ignore_for_file: type=lint
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
+
+part of 'media.dart';
+
+// **************************************************************************
+// FreezedGenerator
+// **************************************************************************
+
+T _$identity<T>(T value) => value;
+
+final _privateConstructorUsedError = UnsupportedError(
+    'It seems like you constructed your class using `MyClass._()`. This constructor is only meant to be used by freezed and you are not supposed to need it nor use it.\nPlease check the documentation here for more information: https://github.com/rrousselGit/freezed#custom-getters-and-methods');
+
+Media _$MediaFromJson(Map<String, dynamic> json) {
+  return _Media.fromJson(json);
+}
+
+/// @nodoc
+mixin _$Media {
+  String get type => throw _privateConstructorUsedError;
+  @JsonKey(name: 'media_keys')
+  List<String> get mediaKeys => throw _privateConstructorUsedError;
+
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+  @JsonKey(ignore: true)
+  $MediaCopyWith<Media> get copyWith => throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $MediaCopyWith<$Res> {
+  factory $MediaCopyWith(Media value, $Res Function(Media) then) =
+      _$MediaCopyWithImpl<$Res>;
+  $Res call({String type, @JsonKey(name: 'media_keys') List<String> mediaKeys});
+}
+
+/// @nodoc
+class _$MediaCopyWithImpl<$Res> implements $MediaCopyWith<$Res> {
+  _$MediaCopyWithImpl(this._value, this._then);
+
+  final Media _value;
+  // ignore: unused_field
+  final $Res Function(Media) _then;
+
+  @override
+  $Res call({
+    Object? type = freezed,
+    Object? mediaKeys = freezed,
+  }) {
+    return _then(_value.copyWith(
+      type: type == freezed
+          ? _value.type
+          : type // ignore: cast_nullable_to_non_nullable
+              as String,
+      mediaKeys: mediaKeys == freezed
+          ? _value.mediaKeys
+          : mediaKeys // ignore: cast_nullable_to_non_nullable
+              as List<String>,
+    ));
+  }
+}
+
+/// @nodoc
+abstract class _$$_MediaCopyWith<$Res> implements $MediaCopyWith<$Res> {
+  factory _$$_MediaCopyWith(_$_Media value, $Res Function(_$_Media) then) =
+      __$$_MediaCopyWithImpl<$Res>;
+  @override
+  $Res call({String type, @JsonKey(name: 'media_keys') List<String> mediaKeys});
+}
+
+/// @nodoc
+class __$$_MediaCopyWithImpl<$Res> extends _$MediaCopyWithImpl<$Res>
+    implements _$$_MediaCopyWith<$Res> {
+  __$$_MediaCopyWithImpl(_$_Media _value, $Res Function(_$_Media) _then)
+      : super(_value, (v) => _then(v as _$_Media));
+
+  @override
+  _$_Media get _value => super._value as _$_Media;
+
+  @override
+  $Res call({
+    Object? type = freezed,
+    Object? mediaKeys = freezed,
+  }) {
+    return _then(_$_Media(
+      type: type == freezed
+          ? _value.type
+          : type // ignore: cast_nullable_to_non_nullable
+              as String,
+      mediaKeys: mediaKeys == freezed
+          ? _value._mediaKeys
+          : mediaKeys // ignore: cast_nullable_to_non_nullable
+              as List<String>,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$_Media implements _Media {
+  const _$_Media(
+      {required this.type,
+      @JsonKey(name: 'media_keys') required final List<String> mediaKeys})
+      : _mediaKeys = mediaKeys;
+
+  factory _$_Media.fromJson(Map<String, dynamic> json) =>
+      _$$_MediaFromJson(json);
+
+  @override
+  final String type;
+  final List<String> _mediaKeys;
+  @override
+  @JsonKey(name: 'media_keys')
+  List<String> get mediaKeys {
+    // ignore: implicit_dynamic_type
+    return EqualUnmodifiableListView(_mediaKeys);
+  }
+
+  @override
+  String toString() {
+    return 'Media(type: $type, mediaKeys: $mediaKeys)';
+  }
+
+  @override
+  bool operator ==(dynamic other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$_Media &&
+            const DeepCollectionEquality().equals(other.type, type) &&
+            const DeepCollectionEquality()
+                .equals(other._mediaKeys, _mediaKeys));
+  }
+
+  @JsonKey(ignore: true)
+  @override
+  int get hashCode => Object.hash(
+      runtimeType,
+      const DeepCollectionEquality().hash(type),
+      const DeepCollectionEquality().hash(_mediaKeys));
+
+  @JsonKey(ignore: true)
+  @override
+  _$$_MediaCopyWith<_$_Media> get copyWith =>
+      __$$_MediaCopyWithImpl<_$_Media>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$_MediaToJson(this);
+  }
+}
+
+abstract class _Media implements Media {
+  const factory _Media(
+          {required final String type,
+          @JsonKey(name: 'media_keys') required final List<String> mediaKeys}) =
+      _$_Media;
+
+  factory _Media.fromJson(Map<String, dynamic> json) = _$_Media.fromJson;
+
+  @override
+  String get type => throw _privateConstructorUsedError;
+  @override
+  @JsonKey(name: 'media_keys')
+  List<String> get mediaKeys => throw _privateConstructorUsedError;
+  @override
+  @JsonKey(ignore: true)
+  _$$_MediaCopyWith<_$_Media> get copyWith =>
+      throw _privateConstructorUsedError;
+}