+++ lib/src/service/tweets/poll.dart
@@ -1,0 +1,22 @@
+// Copyright 2022 Kato Shinya. All rights reserved.
+// Redistribution and use in source and binary forms, with or without
+// modification, are permitted provided the conditions.
+
+// Package imports:
+import 'package:freezed_annotation/freezed_annotation.dart';
+
+// Project imports:
+import 'poll_option.dart';
+
+part 'poll.freezed.dart';
+part 'poll.g.dart';
+
+@freezed
+class Poll with _$Poll {
+  const factory Poll({
+    required String id,
+    required List<PollOption> options,
+  }) = _Poll;
+
+  factory Poll.fromJson(Map<String, Object?> json) => _$PollFromJson(json);
+}